ban hang

+++ DuAn1/src/main/java/domainmodel/SanPham.java
@@ -32,7 +32,7 @@ public class SanPham {
     private String image;
 
     @Column(name = "Gia")
-    private BigDecimal gia;
+    private float gia;
 
     @Column(name = "TrangThai")
     private String trangThai;
@@ -40,7 +40,7 @@ public class SanPham {
     public SanPham() {
     }
 
-    public SanPham(UUID idSP, String tenSP, String moTa, String image, BigDecimal gia, String trangThai) {
+    public SanPham(UUID idSP, String tenSP, String moTa, String image, float gia, String trangThai) {
         this.idSP = idSP;
         this.tenSP = tenSP;
         this.moTa = moTa;
@@ -81,11 +81,11 @@ public class SanPham {
         this.image = image;
     }
 
-    public BigDecimal getGia() {
+    public float getGia() {
         return gia;
     }
 
-    public void setGia(BigDecimal gia) {
+    public void setGia(float gia) {
         this.gia = gia;
     }
